From a open source matlab code
https://uk.mathworks.com/matlabcentral/fileexchange/53070-acoustic-fdtd-example?s_tid=srchtitle_FDTD_4
for sound in air of a room defined in metres (x,y,z). pretty much working in matrices 100% of time. Defines a Gaussian pulse and accounts for pressure, nodes and boundaries. seems like a good starting point in 3D. will have a look at seeing if can implement an adaptive mesh