Finite Element Acoustic Propagation
First version of MATLAB code using the finite element update equations from page 558 in the textbook Computational Ocean Acoustics.
Currently, using the same parameters as the diffuse_FDTD code, the pressure values explode, but reducing the spatial step 'h' does prevent this, however the animation no longer looks like a wave. I've implemented the same visualisation as the diffuse_FDTD code and tried to implement a gaussian pulse source in the centre of the column.