Helmholtz On Disk - MATLAB example
An example provided by MATLAB in the PDE solver documentation, modelling the Helmholtz equation on a disk with a square hole in the centre. A good example of how to define geometry, create a mesh, set coefficients, boundary and initial conditions, solve the PDE and plot the results using MATLAB's predefined functions.